Move all of 'apps' directory into 'web'
