+++ dotcom-rendering/src/web/lib/thrift/nativeConnection.ts
