+++ dotcom-rendering/src/web/lib/thrift/protocols.ts
